move createAllConfigs to separate file
