+++ src/helpers/index.ts
@@ -1,4 +1,5 @@
+import { createAllConfigs } from '@helpers/create-all-configs';
 import { createConfigFile } from '@helpers/create-config-file';
 import { createDirectory } from '@helpers/create-directory';
 
-export { createConfigFile, createDirectory };
+export { createAllConfigs, createConfigFile, createDirectory };
